+++ internal/pg_documentdb_distributed/src/test/regress/sql/bson_aggregate_group_firstlast_sharded.sql
@@ -183,6 +183,7 @@ ROLLBACK;
 
 -- 4b. sortGroup ON: Sort node should disappear, orderby pushed to aggregate
 SET documentdb.enableSortGroupStage TO on;
+SET documentdb.enableSortPushToAccumulator TO on;
 
 set citus.propagate_set_commands to 'local';
 BEGIN;
@@ -194,6 +195,7 @@ SELECT document FROM bson_aggregation_pipeline('db', '{ "aggregate": "fl_sortgro
 ROLLBACK;
 
 RESET documentdb.enableSortGroupStage;
+RESET documentdb.enableSortPushToAccumulator;
 SELECT documentdb_api.drop_collection('db', 'fl_sortgroup_dist');
 
 SET documentdb.enableNewWithExprAccumulators TO off;